+++ NEWS.md
@@ -2,5 +2,9 @@
 
 - Initial Version
 
+# Version 0.2
 
-
+- Bug Fixes on alternative = "less" or "greater"
+- Improved documentation formatting
+- Example data and vignette added
+- wmtype modifications added for Beta methods